requirements.txt
streamlit==1.22.0
streamlit-authenticator==0.4.2
pandas==1.5.3
openpyxl==3.1.2
pyyaml==6.0
jinja2==3.1.2
python-dotenv==1.0.0